+++ features/de.bund.bfr.knime.fsklab.feature/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.4.4] - 2019-01-28
+
 ### Added
 - Trigger Auto complete feature on Arrow down click in Editor JS and Joiner to show all controlled vocabularies associated with this field.
 - Validate the declaration of Output parameters(which are defined in the metadata) in the model script and add warning to the node if they are not so.